Move pool interface to `types` package
