+++ pkg/streamingpromql/operator/instant_vector_selector.go
@@ -15,13 +15,12 @@ import (
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 
-	"github.com/grafana/mimir/pkg/streamingpromql/pooling"
 	"github.com/grafana/mimir/pkg/streamingpromql/types"
 )
 
 type InstantVectorSelector struct {
 	Selector *Selector
-	Pool     pooling.SampleSlicePool
+	Pool     types.SampleSlicePool
 
 	numSteps int
