+++ pkg/streamingpromql/operator/range_vector_function.go
@@ -12,14 +12,13 @@ import (
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/promql"
 
-	"github.com/grafana/mimir/pkg/streamingpromql/pooling"
 	"github.com/grafana/mimir/pkg/streamingpromql/types"
 )
 
 // RangeVectorFunction performs a rate calculation over a range vector.
 type RangeVectorFunction struct {
 	Inner RangeVectorOperator
-	Pool  pooling.SampleSlicePool
+	Pool  types.SampleSlicePool
 
 	numSteps     int
 	rangeSeconds float64
