+++ pkg/streamingpromql/types/data.go
